extract test module
